backend/requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pymongo==4.6.1
motor==3.3.2
redis==5.0.1
influxdb-client==1.39.0

# 数据处理
pandas==2.1.4
numpy==1.26.3
akshare==1.13.16
ta-lib==0.4.28

# 回测框架
backtrader==1.9.78.123

# 任务队列
celery==5.3.6
flower==2.0.1

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
cryptography==42.0.0

# HTTP 客户端
httpx==0.26.0
aiohttp==3.9.1

# 工具库
python-dateutil==2.8.2
pytz==2023.3.post1
loguru==0.7.2

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# 代码质量
black==23.12.1
flake8==7.0.0
mypy==1.8.0
